♻️ Moved optimization_history module from project to optimization pkg
