+++ glotaran/optimization/optimizer.py
@@ -9,10 +9,10 @@ from scipy.optimize import least_squares
 
 from glotaran import __version__ as glotaran_version
 from glotaran.optimization.optimization_group import OptimizationGroup
+from glotaran.optimization.optimization_history import OptimizationHistory
 from glotaran.parameter import ParameterHistory
 from glotaran.project import Result
 from glotaran.project import Scheme
-from glotaran.project.optimization_history import OptimizationHistory
 from glotaran.utils.regex import RegexPattern
 from glotaran.utils.tee import TeeContext
