+++ glotaran/project/result.py
@@ -19,12 +19,12 @@ from glotaran.io import SavingOptions
 from glotaran.io import load_result
 from glotaran.io import save_result
 from glotaran.model import Model
+from glotaran.optimization.optimization_history import OptimizationHistory
 from glotaran.parameter import ParameterGroup
 from glotaran.parameter import ParameterHistory
 from glotaran.project.dataclass_helpers import exclude_from_dict_field
 from glotaran.project.dataclass_helpers import file_loadable_field
 from glotaran.project.dataclass_helpers import init_file_loadable_fields
-from glotaran.project.optimization_history import OptimizationHistory
 from glotaran.project.scheme import Scheme
 from glotaran.utils.io import DatasetMapping
 from glotaran.utils.io import create_clp_guide_dataset
